borrowe route implemented

+++ src/main/java/com/rest_api/fs14backend/serviceImpl/BorrowerServiceImpl.java
@@ -5,6 +5,7 @@ import com.rest_api.fs14backend.entity.BookCopy;
 import com.rest_api.fs14backend.entity.Borrower;
 import com.rest_api.fs14backend.entity.User;
 import com.rest_api.fs14backend.mapper.BorrowMapper;
+import com.rest_api.fs14backend.repository.BookCopyRepository;
 import com.rest_api.fs14backend.repository.BorrowerRepository;
 import com.rest_api.fs14backend.service.BookCopyService;
 import com.rest_api.fs14backend.service.BorrowerService;
@@ -18,8 +19,11 @@ import java.time.LocalDateTime;
 import java.util.Date;
 import java.util.List;
 import java.util.UUID;
+
 @Service
 public class BorrowerServiceImpl implements BorrowerService {
+    @Autowired
+    BookCopyRepository bookCopyRepository;
     @Autowired
     BorrowerRepository borrowerRepository;
     @Autowired
@@ -30,18 +34,25 @@ public class BorrowerServiceImpl implements BorrowerService {
     BorrowMapper borrowMapper;
 
     @Override
-    public Borrower createOne(BorrowDao borrowDao) {
+    public void createOne(BorrowDao borrowDao) {
 
-            UUID userId = borrowDao.getUserId();
-            User user = userService.findById(userId).get();
-            UUID bookCopyId = borrowDao.getBookCopyId();
-            BookCopy bookCopy = bookCopyService.findOne(bookCopyId);
-            // current date as borrow date
-            Date borrowDate = new Date();
-            // add 30 days return time
-            Date returnDate =  new Date(borrowDate.getTime() + ((1000 * 60 * 60 * 24)* 30));
+        UUID userId = borrowDao.getUserId();
+        User user = userService.findById(userId).get();
+        UUID bookCopyId = borrowDao.getBookCopyId();
+        BookCopy bookCopy = bookCopyService.findOne(bookCopyId);
+        // current date as borrow date
+        Date borrowDate = new Date();
+        // add 30 days return time
+        Date returnDate = new Date(borrowDate.getTime() + ((1000 * 60 * 60 * 24) * 30));
+        if (bookCopy != null) {
             Borrower borrower = borrowMapper.toBorrower(user, bookCopy, borrowDate, returnDate);
-            return borrowerRepository.save(borrower);
+            bookCopy.setIsAvailable(false);
+            bookCopyRepository.save(bookCopy);
+            borrowerRepository.save(borrower);
+        } else {
+            throw new RuntimeException("Book is not available");
+        }
+
 
     }
 
@@ -57,7 +68,17 @@ public class BorrowerServiceImpl implements BorrowerService {
 
     @Override
     public void deleteOne(UUID borrowerId) {
-    borrowerRepository.deleteById(borrowerId);
+        Borrower foundBorrower = borrowerRepository.findById(borrowerId).orElse(null);
+        if (foundBorrower != null) {
+            UUID foundBookCopyId = foundBorrower.getBookCopy().getId();
+            BookCopy foundBookCopy = bookCopyRepository.findById(foundBookCopyId).orElse(null);
+
+            if (foundBookCopy != null) {
+                foundBookCopy.setIsAvailable(true);
+                bookCopyRepository.save(foundBookCopy);
+            }
+            borrowerRepository.deleteById(borrowerId);
+        }
     }
